Move form processors to their separate branch.
The form processors are not yet ready for production release.

The problem is how to train them effectively. When gold data is used
when training form processors, results are worse than not using them at
all. The right method is probably to use the network to classify while
training. Nevertheless, because SGD needs to reorder data each
iteration, the form_processors would have to fire after a whole
iteration of training. More research is needed.

For the time being, form processors implementation will reside in
form_processors branch.