Fix compilation on g++-4.9 by refactoring parse_options.
Although {a,b,c} can be used to initialize set, {} does not call
constructor with empty initializer_list, but instead constructor with no
arguments. That is nevertheless explicit for unordered_set in C++11.
(This behaviour is fixed in C++14, where constructor with no arguments
is not marked explicit.)

Nevertheless, using empty set and set with one empty element to denote
no values / any values was abuse, so we now have proper type
option_values with option_values::none and option_values::any
exceptions.